// frontend/src/app/shared/enums/enums.ts
export enum FormMode {
  Add = 'add',
  Edit = 'edit',
  Detail = 'detail'
}

export enum StatusClientId{
  Active = 1,
  Inactive = 2,
  Suspended = 3
}

export enum GenderId{
  Male = 1,
  Female = 2
}
export enum StatusAccountId{
  Enabled = 1,
  Blocked = 2,
  Suspended = 3
}

export enum AccountTypeId{
  Simple = 1,
  Current = 2
}

export enum TransactionTypeId{
  Debit = 1,
  Credit = 2
}